Rename template files
