+++ components/byu-menu/byu-menu.js
@@ -1,5 +1,5 @@
 "use strict";
-import * as template from "./template.html";
+import * as template from "./byu-menu.html";
 import * as util from 'byu-web-component-utils';
 
 class BYUMenu extends HTMLElement {
