+++ components/byu-user-info/byu-user-info.js
@@ -3,7 +3,7 @@
  */
 "use strict";
 
-import * as template from "./template.html";
+import * as template from "./byu-user-info.html";
 import * as util from 'byu-web-component-utils';
 
 class ByuUserInfo extends HTMLElement {
